LCD & hardware changes

+++ database.py
@@ -16,10 +16,11 @@ class Database(Debug, ConfigParser.ConfigParser):
         Debug.__init__(self)
         self.lock = threading.RLock()
         ConfigParser.ConfigParser.__init__(self)
+        test = os.path.join("media", "pi", "data")
         if os.path.isdir(os.path.join("media", "pi", "data")):
             self.__filename = os.path.join("media", "pi", "data", filename)
         else:
-            self.__filename = os.path.join(os.path.basename(__file__), filename)
+            self.__filename = os.path.join(os.path.dirname(__file__), filename)
         if not os.path.exists(self.__filename) and not os.path.exists(self.__filename + ".new"):
             for section in self.SECTIONS:
                 self.add_section(section)
@@ -82,7 +83,8 @@ class Database(Debug, ConfigParser.ConfigParser):
         handle = open(self.__filename + ".new",'w')
         self.write(handle)
         handle.close()
-        os.remove(self.__filename)
+        if os.path.isfile(self.__filename):
+            os.remove(self.__filename)
         os.rename(self.__filename + ".new", self.__filename)
         self.lock.release()
